wmborder file
Changed the behavior from submitting the whole wmborder array to
loading a single npy file again. It turned out that the array in the
memory was copied and thus consumed all the available RAM of the
control node, leading to errors during submission of the SC_row
command. Now all the jobs will load the wmborder from the Harddrive
again.